Assignment 2
Implementing a followup assignment with the original website designed in Assignment 1. The addition of a css style page and several features to make the page more visually pleasing were implemented. All changes were made at the instruction of the "Silly CIA" assignment, except for the addition of the background color. The background color was me. I think it looks nice.